// src/api/generated/api.ts
export interface AttachmentResp {
    'id': number;
    'title': string;
    'link': string;
    'description'?: string;
}